Optimize rs_forney()
Fix a conceptual mistake in rs_forney():
The previous code assumed that the construction of the
error evaluator polynomial should only use the first (2 * n_error)
syndromes. In fact the complete list of syndromes may be used,
but that will not change the result. Just the first (n_error)
syndromes are sufficient to uniquely determine the error values.
So the prevous code was "accidentally correct".

This change optimizes the code further by using the fact that
the degree of the error evaluator polynomial is at most (n_error - 1).